refactor: extract tests
