+++ data/json/items/armor/scarfs_test.ts
@@ -1,0 +1,12 @@
+import { assertEquals } from "https://deno.land/std@0.201.0/assert/assert_equals.ts"
+import scarfs from "./scarfs.json" with { type: "json" }
+import { Item, knit_scarf, knit_scarf_loose, long_knit_scarf, long_knit_scarf_loose, wool_scarf, wool_scarf_loose } from "./scarfs.ts"
+
+const findScarf = ({ id }: Item) => scarfs.find((it) => it.id === id)! as any
+const checkScarf = (scarf: Item, looseScarf: Item) => async (t: Deno.TestContext) => {
+  await t.step(scarf.id, () => assertEquals(scarf, findScarf(scarf)))
+  await t.step(looseScarf.id, () => assertEquals(looseScarf, findScarf(looseScarf)))
+}
+Deno.test("knit_scarf", checkScarf(knit_scarf, knit_scarf_loose))
+Deno.test("long_knit_scarf", checkScarf(long_knit_scarf, long_knit_scarf_loose))
+Deno.test("scarf", checkScarf(wool_scarf, wool_scarf_loose))
